[feature] (stack) Add support to component for CLIENT category
