+++ bigtop-manager-common/src/main/java/org/apache/bigtop/manager/common/utils/shell/DefaultShellResult.java
@@ -1,0 +1,8 @@
+package org.apache.bigtop.manager.common.utils.shell;
+
+import org.apache.bigtop.manager.common.constants.MessageConstants;
+
+public class DefaultShellResult {
+
+    public static final ShellResult SUCCESS = new ShellResult(MessageConstants.SUCCESS_CODE, "Default Successful!!!", "");
+}
