+++ bigtop-manager-stack/bigtop-manager-stack-core/src/main/java/org/apache/bigtop/manager/stack/core/hooks/RestartHookImpl.java
@@ -17,13 +17,11 @@ public class RestartHookImpl implements Hook {
     @Override
     @HookAnnotation(before = HookType.START)
     public void before() {
-        log.info("before restart");
     }
 
     @Override
     @HookAnnotation(after = HookType.START)
     public void after() {
-        log.info("after restart");
     }
 
     @Override
